Fix - Uniformisation des ProductType / details
